+++ fabric/src/main/java/mod/fuji/module/initializer/cleaner/service/CleanerService.java
@@ -62,7 +62,7 @@ public class CleanerService {
         }
 
         if (matcher.getCleanupMethod().equals(CleanupMethod.DISCARD)) {
-            entity.discard();
+            EntityHelper.deleteEntity(entity);
             return;
         }
     }