+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ## ✨ Features
 - [chat.history] New `/chat history forget` command, to forget the remembered chat history.
+- [command_toolbox.delete] New `/delete` command, to delete the target entity without dropping the loot.
 
 ## 🐞 Fixes
 - None